refactor: Change implementation of status of order
Refactor status of order to only allow for one active
order. This might hinder scalability by not allowing
for adding additional states in the future.
However since this is a shopping cart it
might be error prone if the user could have multiple
active orders at the same time. The scalability issue
mentioned could be solved by adding an additional
column in the future if need arises.